order_api/requirements.txt
asgiref==3.5.2
Django==4.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-phonenumber-field==6.3.0
django-phonenumbers==1.0.1
djangorestframework==3.13.1
PyMySQL==1.0.2
pytz==2022.1
sqlparse==0.4.2
